sqlite3: refactors refid handling
Since some members and compounds have a declaration and definition, they'll end
up having more than one memberdef/compounddef that have the same refid. This is
a little tricky to deal with, so I've refactored refids into a struct in order
to store a little information about the context. This also adds some utility
functions for checking whether a member/compound is new or already existed.

It will take several commits for the rest of the code to catch up with this
change.