sqlite3: adds "member" table
- new bi-directional relationship (roughly equivalent to xml listofallmembers)
  that links compounds with members that appear in them.

Just the table/queries for now; implementation will follow in later commits.

Note:
I excluded 2 fields that appear in the XML, "scope" and "name". I'm not 100%
certain on this, but I suspect they are just a convenience feature in the XML;
I haven't seen a scope that is anything but the current namespace, nor a name
that is anything but the name of the memberdef it's pointing at.