sqlite3: adds "rel" view
The rel view, when joined with the ref, compounddef, or memberdef tables,
tacks on an overview of which relations would be fruitful to query. Each column
is just a boolean (1 if any of that relation type exist).

Notes:
1. This may not be the "best" solution, so I'll also state the problem: db users
can't know which relation queries will be fruitful. They either have to pick
relations of interest and always query them, always query all of the relations,
or nevery query any. As I've fleshed out relationship support, this task has
grown significantly.

2. I accidentally committed the tail end of this view in the previous commit
   (525af49bf)