sqlite3: remove some .data() calls
Distinct commit to document something I found a little confusing. As I worked
on this PR, I frequently ran into compile errors like:

error: cannot pass object of non-POD type 'QCString' through variadic function;
call will abort at runtime [-Wnon-pod-varargs]

I don't grok this, so I just parroted an idiom already in the code, and
added .data() calls to many places where a QCString is passed as an argument.

Because I was a little superstitious about this, I made a note to remove all of
these .data() calls when most of the work was done, and then *only keep the ones
that actually block compile*. This got rid of most of the calls I had added, and
most of the ones already in the source.

I ran output database files from before and after this change through `sqldiff`
and they appear to be identical.