declares XMLCodeGenerator in xmlgen.h (1/2)
In order for sqlite3gen to re-use parts of the XML generator, XMLCodeGenerator
needs to be declared in a header file. I parroted how other generators handle
this and have it working in both the XML and sqlite3 outputs, but I don't have
any sense of whether this is otherwise "right".

Diff algo made a confusing history when I tried to make this a single commit,
so I've broken it into two.

This commit:
- cuts the XMLCodeGenerator class out of xmlgen.cpp
- adds the declaration in xmlgen.h

The next commit will re-add the implementation to xmlgen.cpp