sqlite3: fix some 'includes' file names
The sqlite3 generator is failing to link some includes because the "shown" file
path isn't matching up with the full project-relative file path (i.e., it's
recorded as "blah.h", but in the project it's known as "examples/blah.h", so the
naive "blah.h" isn't recognized as matching the known "examples/blah.h" file.

To reduce duplication, I also moved path-prefix-stripping from each call site
into the insertFile method itself (and removed from each call site).

This fix also highlights that we're discarding the "includeName" here (in the
XML, this would be recorded inside the <includes> node). I decided against
adding it for now because the distinction is subtle, potentially confusing, and
I couldn't think of a clear use/need.