sqlite3: doc TODO for output-file safety
I looked into how to check this and don't think I understand the tradeoffs well
enough to come up with a reasonable implementation. Instead I've just documented
it in the hopes that I can get help with this during review, or go hang an issue
report on it after merge.