sqlite3: record group -> member relations
This affects other compounds, but the specific case I had to check
against was group-related.

In the XML output, members get inlined into most surrounding compounds, so it's
trivial to figure out what set of functions belong to the group. In the SQL,
this was only getting formally recorded for ClassDefs via associateAllMembers.

This has a few pieces:
- Discard the member 'ambiguityscope' column. I think this is available on the
  memberdef table (I've added a test to my Python module to verify this)

  This kills a dependency on MemberInfo objects to populate a member relation,
  making it possible to populate member rels from non-ClassDef compounds.
- Refactor associateAllMembers to separate out the logic to associate a single
  member (as associateMember())
- generateSqlite3ForMember uses associateMember to link all members with
  immediate scope definition.
- Rename associateAllMembers to associateAllClassMembers to clarify that its
  logic doesn't apply to other compound types.