sqlite3: abort if database already existed
This could use a quick review.

It attempts to keep the generator from re-populating an existing database,
which produces very noisy output (schema creation and most inserts will fail).

This 'works' on my system, but I have no idea if it is the ideal way to go about
this. I looked around for information on how to check file existence in C++, and
it sounds like this is a complex issue to which the best-answer is a moving
target.